Update footer links and copyright year
Updated the links in the footer section for 'Guide for Authors', 'Ethics and Malpractice Statement' and added a button 'Volumes'. Also changed the footer's 'scroll to top' icon and added an 'indexes' link. The copyright year has been updated to 2023. These changes ensure that the users are directed to the correct web pages and provide better navigation experience.